pyproject.toml
[tool.poetry]
name = "study-django"
version = "0.1.0"
description = ""
authors = ["georgii <goqwertys@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
ipython = "^8.27.0"
django-extensions = "^3.2.3"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.47"
django-countries = "^7.6.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
